Fixed the indexing bug in the SideScrollManager. Updated the ContentManager, SideScrollControls, and added and tested touch events.
Completely rewrote the SideScrollController class and updated all associated classes. Fixed the serious issue with my code - some formulas I'd written for the dragging and sliding were slightly wrong but the effects propagated through much of the component logic. After trying to patch it it was apparent that it would be quicker and cleaner to rewrite some of the logic.