fixed a load of issues:
Rebuilt nav in markup rather than dynamically, fixed nav alignment issues, updated nav text and weights, added media queries in _nav.scss for devices. Added gradient behind nav to lift from content. Changed some of the media queries in content.scss to give more breathing room at the bottom (some still broken though - I think it's more complex than it needs to be). Updated some of the firefly behaviour to fly to the new responsive Alpha nav item position. It does stay locked to that nav item on resize to a certain extent, but it's a bit hacky so needs doing properly as does the whole FireflyManager class when we have time - it will recalculate the nav position each time it flys down to the nav.